SnackBar component scaffolding. This will be the snack bar within the SearchUsers container extracted into a standalone alert component for use throughout the application
