+++ components/layout/Layout.js
@@ -2,7 +2,7 @@ import CssBaseLine from '@material-ui/core/CssBaseline'
 import Container from '@material-ui/core/Container'
 import styled from 'styled-components'
 import MenuDrawer from './MenuDrawer'
-import AppBar from '../appBar/AppBar'
+import AppBar from '../AppBar/AppBar'
 import { useSelector, useDispatch } from 'react-redux'
 import * as Actions from '../../redux/reducers'
 import { useRouter } from 'next/router'
